Fix build system
My build was breaking, and the build system was using some nasty hacks.
I fixed up a lot, and added a Makefile to the root directory.

Building is now a simple ‘make’ or ‘make all’ and cleaning is either
‘make clean’ or ‘make clobber’ (they do the same thing)

There shouldn’t be any warnings or errors when building after this is
all merged in. Thanks :)